fix: modal is not opening properly in user profile
